Windows対応

+++ install.sh
@@ -3,13 +3,22 @@
 cd `dirname $0`
 BASE=${PWD}
 
-unlink ~/.vimrc
-unlink ~/.vim
-unlink ~/.screenrc
+env | grep MSYSTEM_CHOST
 
-ln -s ${BASE}/.vimrc ~/.vimrc
-ln -s ${BASE}/.vim ~/.vim
-ln -s ${BASE}/.screenrc ~/.screenrc
+if [ $? -eq 0 ]; then
+    # win
+    echo "win_mlink.batを管理者権限で実行する"
+    read -p "実行し終わったら何かキーを押してください..."
+else
+    # linux/mac
+    unlink ~/.vimrc
+    unlink ~/.vim
+    unlink ~/.screenrc
+
+    ln -s ${BASE}/.vimrc ~/.vimrc
+    ln -s ${BASE}/.vim ~/.vim
+    ln -s ${BASE}/.screenrc ~/.screenrc
+fi
 
 #++++++++++++++++++++++++++++++
 # .bashrc settings
@@ -23,19 +32,23 @@ fi
 # vim plugin settings
 cd ${BASE}/.vim/plugin
 #alignta.vim
-wget -nc https://raw.githubusercontent.com/h1mesuke/vim-alignta/master/plugin/alignta.vim
+curl -O https://raw.githubusercontent.com/h1mesuke/vim-alignta/master/plugin/alignta.vim
 
 #gtags.vim
-wget -nc http://cvs.savannah.gnu.org/viewvc/*checkout*/global/global/gtags.vim
+curl -O http://cvs.savannah.gnu.org/viewvc/*checkout*/global/global/gtags.vim
 
 #im_control.vim
-wget -nc https://raw.githubusercontent.com/fuenor/im_control.vim/master/plugin/im_control.vim
+curl -O https://raw.githubusercontent.com/fuenor/im_control.vim/master/plugin/im_control.vim
 
 #taglist.vim
-wget -nc https://raw.githubusercontent.com/vim-scripts/taglist.vim/master/plugin/taglist.vim
+curl -O https://raw.githubusercontent.com/vim-scripts/taglist.vim/master/plugin/taglist.vim
 
 #++++++++++++++++++++++++++++++
 # vim color settings
 cd ${BASE}/.vim/colors
 # hybrid
-wget -nc https://raw.githubusercontent.com/w0ng/vim-hybrid/master/colors/hybrid.vim
+curl -O https://raw.githubusercontent.com/w0ng/vim-hybrid/master/colors/hybrid.vim
+
+#++++++++++++++++++++++++++++++
+# execute git clone
+vim
